+++ Project_wsl_2026/Software/main_tennsy4.0/lib/Kicker/kicker.hpp
@@ -7,10 +7,10 @@
 /* キッカー */
 
 // キッカーのピン番号を設定
-void kicker_set_fetpin(int16_t kick_pin, int16_t charge_pin);
+void kicker_set_fetpin(int kick_pin, int charge_pin);
 
 // キッカーのクールダウンの時間を設定・開始
-void kicker_init(int16_t cooldown_time);
+void kicker_init(int cooldown_time);
 
 // キックする（クールダウンタイム中は蹴らない）
 void kicker_kick(bool kick_signal);
@@ -18,7 +18,7 @@ void kicker_kick(bool kick_signal);
 /* キャッチセンサー */
 
 // キャッチセンサーのピン設定・開始
-void catchsensor_init(int16_t pin);
+void catchsensor_init(int pin);
 
 // キャッチセンサーの値を取得
 unsigned int get_catchsensor();
