+++ Project_wsl_2026/Software/main_tennsy4.0/lib/Motors/PD/PDpower.hpp
@@ -5,7 +5,7 @@
 #include "Motors.hpp"
 
 // ジャイロをPD制御で使う場合
-void PD_use_gyro(int16_t target_deg);
+void PD_use_gyro(int target_deg);
 
 // カメラによる角度をPD制御で使う場合
 // 黄色ゴール
@@ -14,6 +14,6 @@ void PD_use_yellow_goal();
 void PD_use_blue_goal();
 
 //  算出したPD出力を取得
-int16_t get_PD_power();
+int get_PD_power();
 
 #endif
